@@ -8,12 +8,14 @@ documentation: https://github.com/chinloyal/pusher_client/blob/master/README.md
 
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0 <2.0.0"
+#  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=2.0.1"
 
 dependencies:
   flutter:
     sdk: flutter
+  build_runner: ^2.0.3
 
 dev_dependencies:
   flutter_test:
