@@ -5,7 +5,7 @@ import 'package:flutter/services.dart';
 abstract class StreamHandler {
   static const EventChannel _eventStream =
       const EventChannel('com.github.chinloyal/pusher_client_stream');
-  StreamSubscription _eventStreamSubscription;
+  StreamSubscription? _eventStreamSubscription;
 
   static Map<String, dynamic Function(dynamic)> _listeners = {};
 
@@ -21,7 +21,7 @@ abstract class StreamHandler {
   /// This method will close the entire event channel stream
   /// which is why it should only be used by [PusherClient]
   void cancelEventChannelStream() {
-    _eventStreamSubscription.cancel();
+    _eventStreamSubscription?.cancel();
   }
 
   void _eventHandler(event) {
