@@ -3,11 +3,11 @@ part 'connection_state_change.g.dart';
 class ConnectionStateChange {
   /// The current connection state. The state the connection has transitioned
   /// to.
-  final String currentState;
+  final String? currentState;
 
   /// The previous connections state. The state the connection has transitioned
   /// from.
-  final String previousState;
+  final String? previousState;
 
   ConnectionStateChange({
     this.currentState,
