@@ -4,17 +4,17 @@ part of 'event_stream_result.dart';
 
 EventStreamResult _$EventStreamResultFromJson(Map<String, dynamic> json) {
   return EventStreamResult(
-    connectionStateChange: json['connectionStateChange'] == null
-        ? null
-        : ConnectionStateChange.fromJson(
-            json['connectionStateChange'] as Map<String, dynamic>),
-    connectionError: json['connectionError'] == null
-        ? null
-        : ConnectionError.fromJson(
-            json['connectionError'] as Map<String, dynamic>),
-    pusherEvent: json['pusherEvent'] == null
-        ? null
-        : PusherEvent.fromJson(json['pusherEvent'] as Map<String, dynamic>),
+    connectionStateChange: json['connectionStateChange'] != null
+        ? ConnectionStateChange.fromJson(
+            json['connectionStateChange'] as Map<String, dynamic>)
+        : null,
+    connectionError: json['connectionError'] != null
+        ? ConnectionError.fromJson(
+            json['connectionError'] as Map<String, dynamic>)
+        : null,
+    pusherEvent: json['pusherEvent'] != null
+        ? PusherEvent.fromJson(json['pusherEvent'] as Map<String, dynamic>)
+        : null,
   );
 }
 
