@@ -5,9 +5,10 @@ part of 'pusher_auth.dart';
 PusherAuth _$PusherAuthFromJson(Map<String, dynamic> json) {
   return PusherAuth(
     json['endpoint'] as String,
-    headers: (json['headers'] as Map<String, dynamic>)?.map(
-      (k, e) => MapEntry(k, e as String),
-    ),
+    headers: json['headers'] != null
+        ? (json['headers'] as Map<String, dynamic>)
+            .map((k, e) => MapEntry(k, e as String))
+        : {},
   );
 }
 
